Приведение к одному стилю выдачи запросов при всех пользователя\фильмах\жанрах и тд. Один тип переменных у моделей. Убраны проверки на нул при создании фильмов.

+++ src/main/java/ru/yandex/practicum/filmorate/storage/reviewsLikes/ReviewLikeStorage.java
@@ -1,11 +1,11 @@
 package ru.yandex.practicum.filmorate.storage.reviewsLikes;
 
 public interface ReviewLikeStorage {
-    void addLike(long reviewId, int userId);
+    void addLike(int reviewId, int userId);
 
-    void addDislike(long reviewId, int userId);
+    void addDislike(int reviewId, int userId);
 
-    boolean deleteLike(long reviewId, int userId);
+    boolean deleteLike(int reviewId, int userId);
 
-    boolean deleteDislike(long reviewId, int userId);
+    boolean deleteDislike(int reviewId, int userId);
 }
